+++ src/components/Link.js
@@ -1,0 +1,7 @@
+// @noflow
+import React from 'react';
+import { Text } from '@quid/react-core';
+
+const Link = props => <Text as="a" type="link bold" {...props} />;
+
+export default Link;